+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -48,11 +48,13 @@ final class DependencyContainer {
   let paywallArchiveManager = PaywallArchiveManager()
 
   init(
+    apiKey: String = "",
     purchaseController controller: PurchaseController? = nil,
     options: SuperwallOptions? = nil
   ) {
     delegateAdapter = SuperwallDelegateAdapter()
     storage = Storage(factory: self)
+    storage.configure(apiKey: apiKey)
     entitlementsInfo = EntitlementsInfo(
       storage: storage,
       delegateAdapter: delegateAdapter
@@ -500,6 +502,14 @@ extension DependencyContainer: StoreTransactionFactory {
   func makeStoreTransaction(from transaction: SK2Transaction) async -> StoreTransaction {
     return StoreTransaction(
       transaction: SK2StoreTransaction(transaction: transaction),
+      configRequestId: configManager.config?.requestId ?? "",
+      appSessionId: appSessionManager.appSession.id
+    )
+  }
+
+  func makeStoreTransaction(from transaction: CustomStoreTransaction) async -> StoreTransaction {
+    return StoreTransaction(
+      transaction: transaction,
       configRequestId: configManager.config?.requestId ?? "",
       appSessionId: appSessionManager.appSession.id
     )